✨ feat: 添加串口接收线程
添加了串口接收线程，为了方便使用，目前该线程使用轮询读取串口数据，以后再考虑优化，并在实际硬件上进行了单元测试。

+++ porting/os/linux/include/os_type.h
@@ -13,7 +13,7 @@ extern "C" {
 #define OS_WAIT_FOREVER (-1)
 
 typedef uint32_t os_time_t;
-typedef int os_stack_t;
+typedef uint8_t os_stack_t;
 
 typedef void *(*os_task_func_t)(void *);
